requirements.txt
aiohttp==3.5.4
aiopg==0.16.0
appdirs==1.4.3
async-timeout==3.0.1
attrs==19.1.0
beautifulsoup4==4.8.0
black==19.3b0
chardet==3.0.4
Click==7.0
idna==2.8
motor==2.0.0
multidict==4.5.2
pkg-resources==0.0.0
psycopg2-binary==2.8.3
pymongo==3.8.0
soupsieve==1.9.2
toml==0.10.0
yarl==1.3.0
sqlalchemy
pyaml